Move metric validation and row key generation to TSUID
