+++ test/core/TestTsdbQueryDownsample.java
@@ -62,7 +62,8 @@ import org.powermock.modules.junit4.PowerMockRunner;
 @PrepareForTest({TSDB.class, Config.class, UniqueId.class, HBaseClient.class,
   CompactionQueue.class, GetRequest.class, PutRequest.class, KeyValue.class,
   Scanner.class, TsdbQuery.class, DeleteRequest.class, Annotation.class,
-  RowKey.class, Span.class, SpanGroup.class, IncomingDataPoints.class })
+  RowKey.class, Span.class, SpanGroup.class, IncomingDataPoints.class,
+  TSUID.class})
 public class TestTsdbQueryDownsample {
 
   private Config config;
@@ -625,7 +626,7 @@ public class TestTsdbQueryDownsample {
     storage = new MockBase(tsdb, client, true, true, true, true);
     storage.setFamily("t".getBytes(MockBase.ASCII()));
 
-    PowerMockito.mockStatic(IncomingDataPoints.class);
+    PowerMockito.mockStatic(TSUID.class);
     PowerMockito.doAnswer(
         new Answer<byte[]>() {
           public byte[] answer(final InvocationOnMock args)
@@ -649,7 +650,7 @@ public class TestTsdbQueryDownsample {
             }
           }
         }
-    ).when(IncomingDataPoints.class, "rowKeyTemplate", (TSDB)any(), anyString(),
+    ).when(TSUID.class, "rowKeyTemplate", (TSDB)any(), anyString(),
         (Map<String, String>)any());
   }
 }
